cmd merge: move all funcs about annotations to annotations.py
